Improve javadocs and OOP
